deleted old store files, and updated ExportMenu
